Rename plotter classes.
