+++ src/einsteinpy/plotting/__init__.py
@@ -1,6 +1,6 @@
-from einsteinpy.plotting.geodesics.geodesics_scatter import ScatterGeodesicPlotter
-from einsteinpy.plotting.geodesics.geodesics_static import StaticGeodesicPlotter
-
-from .mplplotter import MatplotlibPlotter
-from .plotlyplotter import PlotlyPlotter
-from .plotter import Plotter
+from einsteinpy.plotting.geodesics.geodesic_plotter import GeodesicPlotter
+from einsteinpy.plotting.geodesics.interactivegeodesic_plotter import (
+    InteractiveGeodesicPlotter,
+)
+from einsteinpy.plotting.geodesics.staticgeodesic_plotter import StaticGeodesicPlotter
+from einsteinpy.plotting.senile.geodesics_scatter import ScatterGeodesicPlotter
