+++ src/einsteinpy/plotting/senile/geodesics_scatter.py
