+++ src/einsteinpy/tests/test_plotting/test_staticgeodesicplotter.py
@@ -1,10 +1,9 @@
 from unittest import mock
 
 import astropy.units as u
-import matplotlib as mpl
-import matplotlib.pyplot as plt
 import numpy as np
 import pytest
+from matplotlib.axes import Axes
 
 from einsteinpy.bodies import Body
 from einsteinpy.coordinates import SphericalDifferential
@@ -35,30 +34,21 @@ def dummy_data():
 def test_staticgeodesicplotter_has_axes(dummy_data):
     geodesic = dummy_data
     cl = StaticGeodesicPlotter()
-    assert isinstance(cl.ax, mpl.axes.SubplotBase)
-    assert cl.time.value == 0.0
-    assert cl._attractor_present is False
+    assert isinstance(cl.ax, Axes)
+    assert cl.attractor_present is False
 
 
-@mock.patch("einsteinpy.plotting.geodesics.geodesics_static.plt.show")
+@mock.patch("einsteinpy.plotting.geodesics.staticgeodesic_plotter.plt.show")
 def test_plot_calls_plt_show(mock_show, dummy_data):
     geodesic = dummy_data
     cl = StaticGeodesicPlotter()
     cl.plot(geodesic)
     cl.show()
     mock_show.assert_called_with()
-    assert cl._attractor_present
+    assert cl.attractor_present
 
 
-def test_animate_creates_ani(dummy_data):
-    geodesic = dummy_data
-    cl = StaticGeodesicPlotter()
-    cl.animate(geodesic, interval=10)
-    assert cl._attractor_present
-    assert cl.ani
-
-
-@mock.patch("einsteinpy.plotting.geodesics.geodesics_static.plt.savefig")
+@mock.patch("einsteinpy.plotting.geodesics.staticgeodesic_plotter.plt.savefig")
 def test_plot_save_saves_plot(mock_save, dummy_data):
     geodesic = dummy_data
     cl = StaticGeodesicPlotter()
@@ -72,14 +62,12 @@ def test_plot_calls_draw_attractor_Manualscale(dummy_data):
     geodesic = dummy_data
     cl = StaticGeodesicPlotter(attractor_radius_scale=1500)
     cl.plot(geodesic)
-    assert cl._attractor_present
+    assert cl.attractor_present
     assert cl.attractor_radius_scale == 1500
-    assert cl.get_curr_plot_radius != -1
 
 
 def test_plot_calls_draw_attractor_AutoScale(dummy_data):
     geodesic = dummy_data
     cl = StaticGeodesicPlotter()
     cl.plot(geodesic)
-    assert cl._attractor_present
-    assert cl.get_curr_plot_radius != -1
+    assert cl.attractor_present
